+++ CHANGES.rst
@@ -112,6 +112,9 @@ Changelog
   as this is not visible in the application because faceted `sort_on` overrides
   it but it is now taken into account when using restapi with a `base_search_uid`.
   [gbastien]
+- When advice is `asked_again` display left delay correctly, full delay is displayed
+  when advice is supposed given, but when is it `asked_again` it is not the case.
+  [gbastien]
 
 4.2.9b1 (2023-10-27)
 --------------------